refactor: Moved MainMenuBar
